LPD-18212 Base class not needed
