+++ modules/apps/portal-vulcan/portal-vulcan-test/src/testIntegration/java/com/liferay/portal/vulcan/internal/graphql/servlet/test/GraphQLServletTest.java
@@ -6,22 +6,35 @@
 package com.liferay.portal.vulcan.internal.graphql.servlet.test;
 
 import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
+import com.liferay.petra.reflect.ReflectionUtil;
 import com.liferay.petra.string.StringBundler;
 import com.liferay.petra.string.StringPool;
 import com.liferay.portal.kernel.json.JSONArray;
 import com.liferay.portal.kernel.json.JSONObject;
 import com.liferay.portal.kernel.json.JSONUtil;
+import com.liferay.portal.kernel.test.util.HTTPTestUtil;
+import com.liferay.portal.kernel.test.util.RandomTestUtil;
 import com.liferay.portal.kernel.test.util.TestPropsValues;
+import com.liferay.portal.kernel.util.ArrayUtil;
 import com.liferay.portal.kernel.util.HashMapBuilder;
 import com.liferay.portal.kernel.util.HashMapDictionaryBuilder;
+import com.liferay.portal.kernel.util.Http;
 import com.liferay.portal.kernel.util.StringUtil;
 import com.liferay.portal.test.rule.FeatureFlags;
 import com.liferay.portal.test.rule.Inject;
 import com.liferay.portal.test.rule.LiferayIntegrationTestRule;
+import com.liferay.portal.vulcan.graphql.annotation.GraphQLName;
+import com.liferay.portal.vulcan.graphql.annotation.GraphQLTypeExtension;
 import com.liferay.portal.vulcan.graphql.servlet.ServletData;
 import com.liferay.portal.vulcan.internal.test.util.PaginationConfigurationTestUtil;
 
+import java.lang.reflect.Field;
+
+import java.util.Arrays;
 import java.util.Collections;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
 
 import org.junit.After;
 import org.junit.Assert;
@@ -45,7 +58,7 @@ import org.skyscreamer.jsonassert.JSONCompareMode;
  * @author Luis Miguel Barcos
  */
 @RunWith(Arquillian.class)
-public class GraphQLServletTest extends BaseGraphQLServlet {
+public class GraphQLServletTest {
 
 	@ClassRule
 	@Rule
@@ -75,14 +88,14 @@ public class GraphQLServletTest extends BaseGraphQLServlet {
 	public void testMutation() throws Exception {
 		TestDTO testDTO = new TestDTO();
 
-		assertEquals(
+		_assertEquals(
 			false, testDTO,
 			JSONUtil.getValueAsJSONObject(
-				invoke(
+				_invoke(
 					new GraphQLField(
 						"createTestDTO",
 						Collections.singletonMap(
-							"testDTO", toGraphQLString(testDTO)),
+							"testDTO", _toGraphQLString(testDTO)),
 						new GraphQLField("id"), new GraphQLField("map"),
 						new GraphQLField("string")),
 					"mutation"),
@@ -97,16 +110,16 @@ public class GraphQLServletTest extends BaseGraphQLServlet {
 
 		TestDTO testDTO = new TestDTO();
 
-		assertEquals(
+		_assertEquals(
 			false, testDTO,
 			JSONUtil.getValueAsJSONObject(
-				invoke(
+				_invoke(
 					new GraphQLField(
 						"testPath_v1_0",
 						new GraphQLField(
 							"createTestDTO",
 							Collections.singletonMap(
-								"testDTO", toGraphQLString(testDTO)),
+								"testDTO", _toGraphQLString(testDTO)),
 							new GraphQLField("id"), new GraphQLField("map"),
 							new GraphQLField("string"))),
 					"mutation"),
@@ -117,14 +130,14 @@ public class GraphQLServletTest extends BaseGraphQLServlet {
 
 		testDTO = new TestDTO();
 
-		assertEquals(
+		_assertEquals(
 			false, testDTO,
 			JSONUtil.getValueAsJSONObject(
-				invoke(
+				_invoke(
 					new GraphQLField(
 						"createTestDTO",
 						Collections.singletonMap(
-							"testDTO", toGraphQLString(testDTO)),
+							"testDTO", _toGraphQLString(testDTO)),
 						new GraphQLField("id"), new GraphQLField("map"),
 						new GraphQLField("string")),
 					"mutation"),
@@ -133,10 +146,10 @@ public class GraphQLServletTest extends BaseGraphQLServlet {
 
 	@Test
 	public void testQuery() throws Exception {
-		assertEquals(
+		_assertEquals(
 			true, _testDTO,
 			JSONUtil.getValueAsJSONObject(
-				invoke(
+				_invoke(
 					new GraphQLField(
 						"testDTO", new GraphQLField("extendedString"),
 						new GraphQLField("id"), new GraphQLField("map"),
@@ -167,7 +180,7 @@ public class GraphQLServletTest extends BaseGraphQLServlet {
 					"queryDepthLimit", 1
 				).build());
 
-			JSONObject jsonObject = invoke(
+			JSONObject jsonObject = _invoke(
 				new GraphQLField(
 					"testDTO", new GraphQLField("extendedString"),
 					new GraphQLField("id"), new GraphQLField("string")),
@@ -198,10 +211,10 @@ public class GraphQLServletTest extends BaseGraphQLServlet {
 					"queryDepthLimit", 2
 				).build());
 
-			assertEquals(
+			_assertEquals(
 				true, _testDTO,
 				JSONUtil.getValueAsJSONObject(
-					invoke(
+					_invoke(
 						new GraphQLField(
 							"testDTO", new GraphQLField("extendedString"),
 							new GraphQLField("id"), new GraphQLField("map"),
@@ -320,10 +333,10 @@ public class GraphQLServletTest extends BaseGraphQLServlet {
 
 		// With namespace
 
-		assertEquals(
+		_assertEquals(
 			true, _testDTO,
 			JSONUtil.getValueAsJSONObject(
-				invoke(
+				_invoke(
 					new GraphQLField(
 						"testPath_v1_0",
 						new GraphQLField(
@@ -336,10 +349,10 @@ public class GraphQLServletTest extends BaseGraphQLServlet {
 
 		// Without namespace (backwards compatibility)
 
-		assertEquals(
+		_assertEquals(
 			true, _testDTO,
 			JSONUtil.getValueAsJSONObject(
-				invoke(
+				_invoke(
 					new GraphQLField(
 						"testDTO", new GraphQLField("extendedString"),
 						new GraphQLField("id"), new GraphQLField("map"),
@@ -354,7 +367,7 @@ public class GraphQLServletTest extends BaseGraphQLServlet {
 		// Mutation fields
 
 		JSONArray mutationFieldsJSONArray = JSONUtil.getValueAsJSONArray(
-			invoke(
+			_invoke(
 				new GraphQLField(
 					"__schema",
 					new GraphQLField(
@@ -381,7 +394,7 @@ public class GraphQLServletTest extends BaseGraphQLServlet {
 		// Query fields
 
 		JSONArray queryFieldsJSONArray = JSONUtil.getValueAsJSONArray(
-			invoke(
+			_invoke(
 				new GraphQLField(
 					"__schema",
 					new GraphQLField(
@@ -415,7 +428,7 @@ public class GraphQLServletTest extends BaseGraphQLServlet {
 		// Mutation fields
 
 		JSONArray mutationFieldsJSONArray = JSONUtil.getValueAsJSONArray(
-			invoke(
+			_invoke(
 				new GraphQLField(
 					"__schema",
 					new GraphQLField(
@@ -452,7 +465,7 @@ public class GraphQLServletTest extends BaseGraphQLServlet {
 		// Query fields
 
 		JSONArray queryFieldsJSONArray = JSONUtil.getValueAsJSONArray(
-			invoke(
+			_invoke(
 				new GraphQLField(
 					"__schema",
 					new GraphQLField(
@@ -485,6 +498,280 @@ public class GraphQLServletTest extends BaseGraphQLServlet {
 			false, namespacedQueryFieldsJSONArray, false, "testDTO");
 		_assertGraphQLSchemaField(
 			false, namespacedQueryFieldsJSONArray, false, "testDTOPage");
+	}
+
+	public static class TestDTO {
+
+		public TestDTO() {
+			this(
+				RandomTestUtil.randomString(), RandomTestUtil.randomLong(),
+				HashMapBuilder.put(
+					"a" + RandomTestUtil.randomString(),
+					RandomTestUtil.randomString()
+				).put(
+					"a" + RandomTestUtil.randomString(),
+					RandomTestUtil.randomString()
+				).build(),
+				RandomTestUtil.randomString());
+		}
+
+		public TestDTO(
+			String extendedString, long id, Map<String, String> map,
+			String string) {
+
+			_extendedString = extendedString;
+
+			this.id = id;
+			this.map = map;
+			this.string = string;
+		}
+
+		public String getExtendedString() {
+			return _extendedString;
+		}
+
+		public long getId() {
+			return id;
+		}
+
+		public Map<String, String> getMap() {
+			return map;
+		}
+
+		public String getString() {
+			return string;
+		}
+
+		@com.liferay.portal.vulcan.graphql.annotation.GraphQLField
+		protected long id;
+
+		@com.liferay.portal.vulcan.graphql.annotation.GraphQLField
+		protected Map<String, String> map;
+
+		@com.liferay.portal.vulcan.graphql.annotation.GraphQLField
+		protected String string;
+
+		private String _extendedString;
+
+	}
+
+	public static class TestDTOPage {
+
+		public TestDTOPage(int page, int pageSize) {
+			this.page = page;
+			this.pageSize = pageSize;
+		}
+
+		public int getPage() {
+			return page;
+		}
+
+		public int getPageSize() {
+			return pageSize;
+		}
+
+		@com.liferay.portal.vulcan.graphql.annotation.GraphQLField
+		protected int page;
+
+		@com.liferay.portal.vulcan.graphql.annotation.GraphQLField
+		protected int pageSize;
+
+	}
+
+	public static class TestMutation {
+
+		@com.liferay.portal.vulcan.graphql.annotation.GraphQLField
+		public TestDTO createTestDTO(@GraphQLName("testDTO") TestDTO testDTO) {
+			return testDTO;
+		}
+
+	}
+
+	public static class TestQuery {
+
+		public TestQuery() {
+		}
+
+		public TestQuery(TestDTO testDTO) {
+			_testDTO = testDTO;
+		}
+
+		@com.liferay.portal.vulcan.graphql.annotation.GraphQLField
+		public TestDTO testDTO() {
+			return _testDTO;
+		}
+
+		@com.liferay.portal.vulcan.graphql.annotation.GraphQLField
+		public TestDTOPage testDTOPage(
+			@GraphQLName("page") int page,
+			@GraphQLName("pageSize") int pageSize) {
+
+			return new TestDTOPage(page, pageSize);
+		}
+
+		@GraphQLTypeExtension(TestDTO.class)
+		public class TestGraphQLTypeExtension {
+
+			public TestGraphQLTypeExtension(TestDTO testDTO) {
+				_testDTO = testDTO;
+			}
+
+			@com.liferay.portal.vulcan.graphql.annotation.GraphQLField
+			public String extendedString() {
+				return _testDTO.getExtendedString();
+			}
+
+			private final TestDTO _testDTO;
+
+		}
+
+		private static TestDTO _testDTO;
+		private static TestDTOPage _testDTOPage;
+
+	}
+
+	public class GraphQLField {
+
+		public GraphQLField(String key, GraphQLField... graphQLFields) {
+			this(key, new HashMap<>(), graphQLFields);
+		}
+
+		public GraphQLField(
+			String key, Map<String, Object> parameterMap,
+			GraphQLField... graphQLFields) {
+
+			_key = key;
+			_parameterMap = parameterMap;
+			_graphQLFields = Arrays.asList(graphQLFields);
+		}
+
+		@Override
+		public String toString() {
+			StringBuilder sb = new StringBuilder(_key);
+
+			if (!_parameterMap.isEmpty()) {
+				sb.append("(");
+
+				for (Map.Entry<String, Object> entry :
+						_parameterMap.entrySet()) {
+
+					sb.append(entry.getKey());
+					sb.append(": ");
+					sb.append(entry.getValue());
+					sb.append(", ");
+				}
+
+				sb.setLength(sb.length() - 2);
+
+				sb.append(")");
+			}
+
+			if (!_graphQLFields.isEmpty()) {
+				sb.append("{");
+
+				for (GraphQLField graphQLField : _graphQLFields) {
+					sb.append(graphQLField.toString());
+					sb.append(", ");
+				}
+
+				sb.setLength(sb.length() - 2);
+
+				sb.append("}");
+			}
+
+			return sb.toString();
+		}
+
+		private final List<GraphQLField> _graphQLFields;
+		private final String _key;
+		private final Map<String, Object> _parameterMap;
+
+	}
+
+	public class TestServletData implements ServletData {
+
+		public TestServletData(TestDTO testDTO) {
+			_testQuery = new TestQuery(testDTO);
+		}
+
+		@Override
+		public String getApplicationName() {
+			return "test";
+		}
+
+		@Override
+		public Object getMutation() {
+			return _testMutation;
+		}
+
+		@Override
+		public String getPath() {
+			return "/test-path-graphql/v1_0";
+		}
+
+		@Override
+		public TestQuery getQuery() {
+			return _testQuery;
+		}
+
+		@Override
+		public boolean isJaxRsResourceInvocation() {
+			return false;
+		}
+
+		private final TestMutation _testMutation = new TestMutation();
+		private final TestQuery _testQuery;
+
+	}
+
+	private void _appendGraphQLFieldValue(StringBuilder sb, Object value) {
+		if (value instanceof Map) {
+			Map<String, Object> map = (Map)value;
+
+			sb.append("{");
+
+			StringBuilder stringBuilder = new StringBuilder();
+
+			for (Map.Entry<String, Object> entry : map.entrySet()) {
+				if (stringBuilder.length() > 1) {
+					stringBuilder.append(", ");
+				}
+
+				stringBuilder.append(entry.getKey());
+				stringBuilder.append(": ");
+
+				_appendGraphQLFieldValue(stringBuilder, entry.getValue());
+			}
+
+			sb.append(stringBuilder.toString());
+			sb.append("}");
+		}
+		else if (value instanceof String) {
+			sb.append("\"");
+			sb.append(value);
+			sb.append("\"");
+		}
+		else {
+			sb.append(value);
+		}
+	}
+
+	private void _assertEquals(
+		boolean assertExtendedProperties, TestDTO expectedTestDTO,
+		JSONObject jsonObject) {
+
+		if (assertExtendedProperties) {
+			Assert.assertEquals(
+				expectedTestDTO.getExtendedString(),
+				jsonObject.get("extendedString"));
+		}
+
+		Assert.assertEquals(expectedTestDTO.getId(), jsonObject.get("id"));
+		Assert.assertEquals(
+			expectedTestDTO.getMap(),
+			JSONUtil.toStringMap(jsonObject.getJSONObject("map")));
+		Assert.assertEquals(
+			expectedTestDTO.getString(), jsonObject.get("string"));
 	}
 
 	private void _assertGraphQLSchemaField(
@@ -546,13 +833,25 @@ public class GraphQLServletTest extends BaseGraphQLServlet {
 		return null;
 	}
 
+	private JSONObject _invoke(GraphQLField graphQLField, String type)
+		throws Exception {
+
+		GraphQLField queryGraphQLField = new GraphQLField(type, graphQLField);
+
+		return HTTPTestUtil.invokeToJSONObject(
+			JSONUtil.put(
+				"query", queryGraphQLField.toString()
+			).toString(),
+			"graphql", Http.Method.POST);
+	}
+
 	private void _test(
 			int expectedPage, int expectedPageSize, Integer requestPage,
 			Integer requestPageSize)
 		throws Exception {
 
 		JSONObject jsonObject = JSONUtil.getValueAsJSONObject(
-			invoke(
+			_invoke(
 				new GraphQLField(
 					"testDTOPage",
 					HashMapBuilder.put(
@@ -566,6 +865,33 @@ public class GraphQLServletTest extends BaseGraphQLServlet {
 
 		Assert.assertEquals(expectedPage, jsonObject.getInt("page"));
 		Assert.assertEquals(expectedPageSize, jsonObject.getInt("pageSize"));
+	}
+
+	private String _toGraphQLString(TestDTO testDTO) throws Exception {
+		StringBuilder sb = new StringBuilder("{");
+
+		for (Field field : ReflectionUtil.getDeclaredFields(TestDTO.class)) {
+			if (ArrayUtil.isEmpty(
+					field.getAnnotationsByType(
+						com.liferay.portal.vulcan.graphql.annotation.
+							GraphQLField.class))) {
+
+				continue;
+			}
+
+			if (sb.length() > 1) {
+				sb.append(", ");
+			}
+
+			sb.append(field.getName());
+			sb.append(": ");
+
+			_appendGraphQLFieldValue(sb, field.get(testDTO));
+		}
+
+		sb.append("}");
+
+		return sb.toString();
 	}
 
 	@Inject
